pyproject.toml
[tool.poetry]
name = "image-service"
version = "0.1.0"
description = "service to serve images"
authors = ["Mark Antipin <antipinsuperstar@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
backend-utils = "^0.2.0"
python-dotenv = "^0.21.0"
aioboto3 = "^10.1.0"
uvicorn = "^0.19.0"
python-multipart = "^0.0.5"
Pillow = "^9.2.0"
aiobotocore = "^2.4.0"
aiofiles = "^22.1.0"
asynclog = "^0.1.7"
starlette-exporter = "^0.14.0"
opencv-python = "^4.6.0"
python-magic = "^0.4.27"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
flake8-import-order = "^0.18.1"
flake8-bugbear = "^22.10.27"
flake8-eradicate = "^1.4.0"
flake8-builtins = "^2.0.0"
flake8-comprehensions = "^3.10.0"
flake8-broken-line = "^0.6.0"
flake8-print = "^5.0.0"
pep8-naming = "^0.13.2"
flake8-unused-arguments = "^0.0.12"
mypy = "^0.982"
isort = "^5.10.1"
requests = "^2.28.1"
pytest-mock = "^3.10.0"
coverage = "^6.5.0"
testfixtures = "^7.0.3"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
